pyproject.toml
[tool.poetry]
name = "ai-worker"
version = "0.3.0"
description = "An ai worker that posts capabilities, and accepts jobs"
authors = ["erik aronesty <erik@q32.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ai_worker"}, {include = "gguf_loader"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
llama-cpp-python = {extras = ["server"], version = "0.2.7"}
pydantic-settings = "^2.0.3"
fastapi = "^0.103.1"
websockets = "^11.0.3"
httpx = "^0.24.1"
psutil = "^5.9.5"
huggingface-hub = "^0.16.4"
sentencepiece = "^0.1.99"
python-dotenv = "^1.0.0"
pynvml = "^11.5.0"
gguf = "^0.3.2"
itsdangerous = "^2.1.2"
httpx-sse = "^0.3.1"
pyopencl = "^2023.1.2"
coincurve = "^18.0.0"
# https://github.com/huggingface/datasets/issues/6330
fsspec = "2023.9.2"

[tool.poetry.group.torch]
optional = true

[tool.poetry.group.torch.dependencies]
bitsandbytes = "^0.41.1"
torch = ">=2.1.0"
datasets = "^2.14.5"
scipy = "^1.11.3"
transformers = "*"
accelerate = "^0.23.0"
peft = "^0.5.0"
cmake = "^3.27.7"
lit = "^17.0.3"
nvidia-cublas-cu12=">=2.1.3.1"
nvidia-cuda-cupti-cu12=">=12.1.105"
nvidia-cuda-nvrtc-cu12=">=12.1.105"
nvidia-cuda-runtime-cu12=">=12.1.105"
nvidia-cudnn-cu12="8.9.4.25"
nvidia-cufft-cu12=">=11.0.2.54"
nvidia-curand-cu12=">=10.3.2.106"
nvidia-cusolver-cu12=">=11.4.5.107"
nvidia-cusparse-cu12=">=12.1.0.106"
nvidia-nccl-cu12=">=2.18.1"
nvidia-nvjitlink-cu12=">=12.3.52"
nvidia-nvtx-cu12=">=12.1.105"
triton = ">=2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pyinstaller = "^5.13.2"
ruff = "^0.0.288"
toml = "^0.10.2"
coverage = "^7.3.2"


[tool.poetry.group.onnx]
optional = true


[tool.poetry.group.onnx.dependencies]
onnx = "^1.15.0"
tqdm = "^4.66.1"
onnxruntime-gpu = "^1.16.1"
fastembed = "^0.1.1"
optimum = {extras = ["onnxruntime-gpu"], version = "^1.13.2"}
diffusers = "^0.21.4"
Pillow = "^10.1.0"


[tool.poetry.group.dev-onnx]
optional = true


[tool.poetry.group.dev-onnx.dependencies]
onnx = "^1.15.0"
tqdm = "^4.66.1"
fastembed = "^0.1.1"
onnxruntime = "^1.16.1"
optimum = {extras = ["onnxruntime-gpu"], version = "^1.13.2"}
diffusers = "^0.21.4"
Pillow = "^10.1.0"

[tool.ruff]
line-length = 120
ignore = ['E701', 'E501', 'E402']

[tool.ruff.per-file-ignores]
"__init__.py" = ['F401']

[tool.poetry.scripts]
gputopia_worker = 'ai_worker.main:main'
quantize_gguf = 'ai_worker.quantize_main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
